Interaction WORKS!!!!!
holy BALLS this took me way longer than it should have but it's also the first thing I've done basically from scratch in unity so go me. I need to put together clean looking textboxes and then we should be golden. For texts with multiple pages, we can set a variable that marks off each page. Each time the submit is pressed, the interaction occurs and the number is incremented. The page is destroyed and the player released after the last page is pressed. This will also allow us to make decisions, cool beans